ok configuration prix

+++ vol/src/main/java/com/itu/vol/repository/VolRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VolRepository extends JpaRepository<Vol, Long> {
+public interface VolRepository extends JpaRepository<Vol, Integer> {
 }
